pyproject.toml
[project]
name = "helpers-system-api"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.0",
    "fastapi[standard]>=0.135.2",
    "greenlet>=3.2.4",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.41",
    "asyncpg>=0.30.0",
]

[dependency-groups]
dev = [
    "anyio>=4.13.0",
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pydantic[mypy]>=2.12.5",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.8",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
addopts = "--cov=app --cov-report=term-missing --cov-report=xml"
markers = ["integration: tests that require a live database"]
filterwarnings = ["ignore::jwt.warnings.InsecureKeyLengthWarning"]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.14"
strict = true
plugins = ["pydantic.mypy"]
exclude = ["alembic/"]

[[tool.mypy.overrides]]
module = ["jwt.*", "redis.*"]
ignore_missing_imports = true
